backend/requirements.txt
asgiref==3.8.1
backports.zoneinfo==0.2.1
Django==4.2.16
django-cors-headers==4.4.0
djangorestframework==3.15.2
djangorestframework-simplejwt==5.3.1
pillow==10.4.0
PyJWT==2.9.0
PyMySQL==1.1.1
sqlparse==0.5.2
typing-extensions==4.12.2
whitenoise==6.7.0
